move getSession logic to login-library
Jira ticket: CAMS-401
